fix: better design for sip to make it more reliable @todo CallLifecycle and Session.SetState() merge
