+++ api/assistant-api/internal/channel/telephony/internal/sip/streamer.go
@@ -38,8 +38,7 @@ type Streamer struct {
 	ringbackCancel      context.CancelFunc
 	onTransferInitiated func(targets []string, message string, postTransferAction string)
 
-	ctx    context.Context
-	cancel context.CancelFunc
+	cancelParent context.CancelFunc
 }
 
 func NewStreamer(ctx context.Context,
@@ -51,7 +50,7 @@ func NewStreamer(ctx context.Context,
 	if sipSession == nil {
 		return nil, fmt.Errorf("SIP session is required — standalone server mode is not supported")
 	}
-	streamerCtx, cancel := context.WithCancel(ctx)
+	_, cancel := context.WithCancel(ctx)
 
 	s := &Streamer{
 		BaseTelephonyStreamer: internal_telephony_base.NewBaseTelephonyStreamer(
@@ -59,21 +58,22 @@ func NewStreamer(ctx context.Context,
 			internal_telephony_base.WithSourceAudioConfig(internal_audio.NewMulaw8khzMonoAudioConfig()),
 			internal_telephony_base.WithBaseOption(channel_base.WithInputAudioConfig(internal_audio.NewLinear16khzMonoAudioConfig())),
 		),
-		ctx:    streamerCtx,
-		cancel: cancel,
+		cancelParent: cancel,
 	}
 
 	go func() {
-		<-streamerCtx.Done()
-		reason := protos.ConversationDisconnection_DISCONNECTION_TYPE_UNSPECIFIED
 		select {
 		case <-sipSession.ByeReceived():
-			reason = protos.ConversationDisconnection_DISCONNECTION_TYPE_USER
-		default:
+			s.Logger.Infow("SIP streamer: user BYE received")
+		case <-sipSession.Context().Done():
+			s.Logger.Infow("SIP streamer: session context cancelled")
+		case <-s.Ctx.Done():
+			return
 		}
-		if msg := s.Disconnect(reason); msg != nil {
+		if msg := s.Disconnect(protos.ConversationDisconnection_DISCONNECTION_TYPE_USER); msg != nil {
 			s.Input(msg)
 		}
+		s.Close()
 	}()
 
 	rtpHandler := sipSession.GetRTPHandler()
@@ -91,8 +91,8 @@ func NewStreamer(ctx context.Context,
 	})
 
 	go s.forwardIncomingAudio()
-	go s.audio.RunOutputSender(streamerCtx)
-	go s.audio.RunBridgeRecorder(streamerCtx)
+	go s.audio.RunOutputSender(s.Ctx)
+	go s.audio.RunBridgeRecorder(s.Ctx)
 	s.Input(s.CreateConnectionRequest())
 
 	localIP, localPort := rtpHandler.LocalAddr()
@@ -119,7 +119,7 @@ func (s *Streamer) forwardIncomingAudio() {
 	bufferThreshold := s.InputBufferThreshold()
 	for {
 		select {
-		case <-s.ctx.Done():
+		case <-s.Ctx.Done():
 			return
 		case audioData, ok := <-rtpHandler.AudioIn():
 			if !ok {
@@ -157,7 +157,7 @@ func (s *Streamer) forwardIncomingAudio() {
 }
 
 func (s *Streamer) Context() context.Context {
-	return s.ctx
+	return s.Ctx
 }
 
 func (s *Streamer) Send(response internal_type.Stream) error {
@@ -175,10 +175,12 @@ func (s *Streamer) Send(response internal_type.Stream) error {
 			s.audio.ClearOutputBuffer()
 		}
 	case *protos.ConversationDisconnection:
+		s.Logger.Infow("SIP streamer: Send(ConversationDisconnection)", "type", data.GetType().String())
 		if disc := s.Disconnect(data.GetType()); disc != nil {
 			s.Input(disc)
 		}
 		s.endSession()
+		s.Close()
 	case *protos.ConversationToolCall:
 		switch data.GetAction() {
 		case protos.ToolCallAction_TOOL_CALL_ACTION_END_CONVERSATION:
@@ -229,16 +231,14 @@ func (s *Streamer) EnterTransferMode(targets []string, message string, postTrans
 		session.SetState(sip_infra.CallStateTransferring)
 	}
 
-	// If a transfer message was provided, TTS is still playing — let it finish.
-	// Otherwise clear the AI output and play a ringback tone.
-	if message == "" {
-		s.audio.ClearOutputBuffer()
-		ringbackCtx, ringbackCancel := context.WithCancel(s.ctx)
-		s.mu.Lock()
-		s.ringbackCancel = ringbackCancel
-		s.mu.Unlock()
-		go s.audio.PlayRingback(ringbackCtx)
-	}
+	ringbackCtx, ringbackCancel := context.WithCancel(s.Ctx)
+	s.mu.Lock()
+	s.ringbackCancel = ringbackCancel
+	s.mu.Unlock()
+	go func() {
+		s.audio.WaitOutputDrain(ringbackCtx)
+		s.audio.PlayRingback(ringbackCtx)
+	}()
 
 	if callback != nil {
 		callback(targets, message, postTransferAction)
@@ -316,7 +316,7 @@ func (s *Streamer) endSession() {
 	s.mu.RLock()
 	session := s.session
 	s.mu.RUnlock()
-	if session != nil && !s.transferring.Load() {
+	if session != nil {
 		session.End()
 	}
 }
@@ -330,17 +330,13 @@ func (s *Streamer) Close() error {
 		return nil
 	}
 
-	s.cancel()
+	s.cancelParent()
 	s.BaseStreamer.Cancel()
 	s.ResetInputBuffer()
 
 	s.mu.RLock()
 	session := s.session
 	s.mu.RUnlock()
-
-	if s.transferring.Load() {
-		return nil
-	}
 
 	if session != nil {
 		session.End()
